Optim

+++ test/test_lc.py
@@ -10,19 +10,18 @@ from pypowspec import compute_auto_lc, compute_cross_lc
 import pandas as pd
 dat_fname = "/global/project/projectdirs/desi/mocks/UNIT/HOD_Shadab/multiple_snapshot_lightcone/UNIT_lightcone_multibox_ELG_footprint_nz_NGC.dat"
 ran_fname = "/global/project/projectdirs/desi/mocks/UNIT/HOD_Shadab/multiple_snapshot_lightcone/UNIT_lightcone_multibox_ELG_footprint_nz_1xdata_5.ran_NGC.dat"
-data = pd.read_csv(dat_fname, usecols = (0,1,3,4), engine='c', delim_whitespace=True, names = ['ra', 'dec', 'zrsd', 'nz']).values
-rand = pd.read_csv(ran_fname, usecols = (0,1,3,4), engine='c', delim_whitespace=True, names = ['ra', 'dec', 'zrsd', 'nz']).values
+data = pd.read_csv(dat_fname, usecols = (0,1,3,4), engine='c', delim_whitespace=True, names = ['ra', 'dec', 'zrsd', 'nz']).values.astype(np.float32)
+rand = pd.read_csv(ran_fname, usecols = (0,1,3,4), engine='c', delim_whitespace=True, names = ['ra', 'dec', 'zrsd', 'nz']).values.astype(np.float32)
 
 data = data[(data[:,2] > 0.7) & (data[:,2] < 1.)]
 rand = rand[(rand[:,2] > 0.7) & (rand[:,2] < 1.)]
 
-fkp_data = 1. / (1 + 5000 * data[:,3])
-fkp_rand = 1. / (1 + 5000 * rand[:,3])
-
+fkp_data = 1. / (1 + 5000 * data[:,3]).astype(np.float32)
+fkp_rand = 1. / (1 + 5000 * rand[:,3]).astype(np.float32)
 nobj = data.shape[0]
-wdata = np.ones(nobj)
+wdata = np.ones(nobj).astype(np.float32)
 
-wrand = np.ones(rand.shape[0])
+wrand = np.ones(rand.shape[0]).astype(np.float32)
